rime-schema-compare/requirements.txt
pypinyin>=0.50.0
python-dotenv>=1.0.0
PyYAML>=6.0
